ai-resume-job-matching/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
sentence-transformers==5.1.2
pandas
python-multipart
python-dotenv==1.2.1
pydantic==2.12.5
langchain==0.3.28
langchain-openai==0.3.35
langchain-core==0.3.83
langchain-text-splitters==0.3.11
openai==2.28.0
bcrypt==4.1.3
python-jose[cryptography]==3.5.0
SQLAlchemy==2.0.48
google-auth==2.49.1
chromadb
pypdf
endee
